Use a local generator if specified by CSHARP_GENERATOR_DIR

+++ CSharpGeneratorFunctions.sh
@@ -1,18 +1,28 @@
 # This is intended to be imported using the "source" function
 # from any script that wants to invoke the C#-based generator.
 
-CSHARP_GENERATOR_DIR=tmp-gapic-generator-csharp
+TMP_CSHARP_GENERATOR_DIR=tmp-gapic-generator-csharp
+if [[ $CSHARP_GENERATOR_DIR == "" ]]
+then
+  CSHARP_GENERATOR_DIR=$TMP_CSHARP_GENERATOR_DIR
+fi
 
-# Installs the C# generator in a tmp-gapic-generator-csharp directory.
-# If the directory already exists, it is deleted first.
+# Installs the C# generator in a tmp-gapic-generator-csharp directory,
+# unless it an existing generator directory is specified via the
+# CSHARP_GENERATOR_DIR environment variable.
+# When installing, if the directory already exists, it is deleted first.
 install_csharp_generator() {
-  echo "Installing C# generator in $CSHARP_GENERATOR_DIR"
-  rm -rf $CSHARP_GENERATOR_DIR
+  if [[ $CSHARP_GENERATOR_DIR == $TMP_CSHARP_GENERATOR_DIR ]]
+  then
+    echo "Installing C# generator in $CSHARP_GENERATOR_DIR"
+    rm -rf $CSHARP_GENERATOR_DIR
   
-  git clone --quiet --recursive --depth=1 \
-    https://github.com/googleapis/gapic-generator-csharp.git \
-    $CSHARP_GENERATOR_DIR
-    
+    git clone --quiet --recursive --depth=1 \
+      https://github.com/googleapis/gapic-generator-csharp.git \
+      $CSHARP_GENERATOR_DIR
+  else
+    echo "Using C# generator in $CSHARP_GENERATOR_DIR"
+  fi
   echo 'Building C# generator'
   (cd $CSHARP_GENERATOR_DIR && \
    dotnet build -c Release -nologo -clp:NoSummary -v quiet Google.Api.Generator.Rest)
